SmartNews-AI-main/AiNewsProject/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy[asyncio]==2.0.23
aiomysql==0.2.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Environment Variables
python-dotenv==1.0.0

# File Upload
python-multipart==0.0.6

# CORS
starlette==0.27.0

# Pydantic (for data validation)
pydantic==2.5.0
pydantic-settings==2.1.0

# redis
redis==7.1.0
